file intake and output methods from DataManipulator class have been moved to a new class named DataInputOutput
