+++ sources/common/DataInputOutput_Locale/en_US/LC_MESSAGES/DataInputOutput.po
@@ -1,0 +1,27 @@
+# Copyright (C) 2020 Daniel Popiniuc
+# This file is distributed under the same license as the DataInputOutput package.
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: DataInputOutput\n"
+"POT-Creation-Date: 2020-04-25 13:20+0200\n"
+"PO-Revision-Date: 2020-05-02 01:16+0300\n"
+"Last-Translator: danielppiniuc@gmail.com\n"
+"Language-Team: English\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: en_US\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+
+msgid "All relevant files ({files_counted}) were merged into a Pandas Data Frame"
+msgstr "All relevant files ({files_counted}) were merged into a Pandas Data Frame"
+
+msgid "File \"format\" attribute has a value of \"{format_value}\" which is not among currently implemented values: \"{implemented_file_formats}\", therefore file saving is not possible"
+msgstr "File \"format\" attribute has a value of \"{format_value}\" which is not among currently implemented values: \"{implemented_file_formats}\", therefore file saving is not possible"
+
+msgid "File \"format\" attribute is mandatory in the file setting, but missing, therefore file saving is not possible"
+msgstr "File \"format\" attribute is mandatory in the file setting, but missing, therefore file saving is not possible"
+
+msgid "Pandas Data Frame has just been saved to file \"{file_name}\", considering {file_type} as file type"
+msgstr "Pandas Data Frame has just been saved to file \"{file_name}\", considering {file_type} as file type"
