+++ sources/common/DataInputOutput_Locale/it_IT/LC_MESSAGES/DataInputOutput.po
@@ -1,0 +1,27 @@
+# Copyright (C) 2020 Daniel Popiniuc
+# This file is distributed under the same license as the DataInputOutput package.
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: DataInputOutput\n"
+"POT-Creation-Date: 2020-04-25 13:20+0200\n"
+"PO-Revision-Date: 2020-05-02 01:16+0300\n"
+"Last-Translator: danielppiniuc@gmail.com\n"
+"Language-Team: Italian\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: it\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+
+msgid "All relevant files ({files_counted}) were merged into a Pandas Data Frame"
+msgstr "Tutti i file rilevanti ({files_counted}) sono stati uniti in una Strutura di Dati Pandas"
+
+msgid "File \"format\" attribute has a value of \"{format_value}\" which is not among currently implemented values: \"{implemented_file_formats}\", therefore file saving is not possible"
+msgstr "L'attributo file \"format\" ha un valore di \"{format_value}\" che non è tra i valori attualmente implementati: \"{implemented_file_formats}\", quindi il salvataggio del file non è possibile"
+
+msgid "File \"format\" attribute is mandatory in the file setting, but missing, therefore file saving is not possible"
+msgstr "L'attributo file \"formato\" è obbligatorio nell'impostazione del file, ma manca, pertanto non è possibile salvare il file"
+
+msgid "Pandas Data Frame has just been saved to file \"{file_name}\", considering {file_type} as file type"
+msgstr "La Strutura di Dati Pandas è stato appena salvato nel file \"{file_name}\", considerando {file_type} come tipo di file"
