+++ sources/common/DataInputOutput_Locale/ro_RO/LC_MESSAGES/DataInputOutput.po
@@ -1,0 +1,27 @@
+# Copyright (C) 2020 Daniel Popiniuc
+# This file is distributed under the same license as the DataInputOutput package.
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: DataInputOutput\n"
+"POT-Creation-Date: 2020-04-25 13:20+0200\n"
+"PO-Revision-Date: 2020-05-02 01:16+0300\n"
+"Last-Translator: danielppiniuc@gmail.com\n"
+"Language-Team: Romanian\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: ro\n"
+"Plural-Forms: nplurals=3; plural=n==1 ? 0 : (n==0 || (n%100 > 0 && n%100 < 20)) ? 1 : 2;\n"
+
+msgid "All relevant files ({files_counted}) were merged into a Pandas Data Frame"
+msgstr "Toate fișierele relevante ({files_counted}) au fost contopite într-un Cadru de Date Pandas"
+
+msgid "File \"format\" attribute has a value of \"{format_value}\" which is not among currently implemented values: \"{implemented_file_formats}\", therefore file saving is not possible"
+msgstr "Atributul \"format\" al fișierului are o valoare \"{format_value}\" care nu se află printre valorile implementate în prezent: \"{implemented_file_formats}\", prin urmare, salvarea fișierului nu este posibilă"
+
+msgid "File \"format\" attribute is mandatory in the file setting, but missing, therefore file saving is not possible"
+msgstr "Atributul \"format\" al fișierului este obligatoriu, dar acesta lipsește, prin urmare nu este posibilă salvarea fișierului"
+
+msgid "Pandas Data Frame has just been saved to file \"{file_name}\", considering {file_type} as file type"
+msgstr "Cadru de Date Pandas tocmai a fost salvat în fișierul \"{file_name}\", considerând {file_type} ca tip de fișier"
